Replace callcabal2nix with cabal2nix
callcabal2nix uses IFD, which means we can't evaluate this package for
systems other than our own without something like
`boot.binfmt.emulatedSystems = [ ... ]`.

It's a bit sad that we will need to manually update default.nix, but I
think given that this library is fairly stable that is an ok tradeoff to
make to avoid IFD in the exported packages.